allow several projects / organisations + started to incorporate timestamps
I now allow that several projects or organisations are given in the job
string. Format is that of people, i.e.: A (,|and) B (,|and) ... . I had
setup the database already in that way, but forgot about it in the
meantime.

I also moved towards handling timestamps. I have decided to use python's
built-in converters to store timestamps in the sqlite-DB.