default activities for projects + bugfixes
I noticed that I sometimes used words as activities and as project names at the
same time, e.g., I have job strings like 'reading for DySCo' which is reading I
did for a meeting and then 'DySCo' which is attending the meeting. I now
introduced a way of handling this in ActivityLog: Activity strings may now also
be project strings (change in parseJobStr). If a project is used in place of an
activity for the first time, the user is asked to provide a default activity for
that project. For example, 'DySCo' would get 'meeting' as default activity.
Whenever the project is used without explicit activity the default activity is
logged with the job. I have another example where I used this: 'admin'. Here
there is no clear activity associated with a potential project 'admin'. 'admin'
could also stand for administrational stuff which could be an activity and not a
project, but I decided that it's better to stay consistent and define it as
project as well. Then it is possible to assign an organisation to it. Although
that doesn't matter, because you can assign organisations to jobs, too. Anyway,
as 'admin' could be any type of activity which I don't want to specify further
(e.g., reading, writing, filling out forms, asking how something is done,
signing stuff, ordering), I added an unspecified 'mixed' default activity which
can now be assigned as a default activity to a project.

bugfixes:
- logical error in getTime
- re-introduced case sensitivity for activities (to make compatible with
  projects)