First version of the ActivityLog filler function
This takes a semicolon-separated file containing a list of log entries converted
from Matlab and puts the entries in an ActivityLog. The standard functions of
the ActivityLog are used and the user has to clarify input that is unclear, as
when running the ActivityLog directly.

The filler breaks on errors and resumes work when called again. This is
implemented by storing the last successful state (the last row) in
lastrowinfo.pkl. The user can also break the process by pressing Ctrl-c followed
by Enter in one of the user input dialogues. Filling will then also be resumed
the next time the filler is called.

Seems to work.