bugfix + copied Matlab logs into oldlogs
The converter did not handle negative durations well: I forgot that there are
two possible cases, one in which the start time of the job with negative
duration was to blame and one in which the end time of the job with negative
duration was to blame. The user has to resolve this now.

I have copied all logfiles from Matlab into directory 'oldlogs'.