check whether duration of last job present to handle crashes
I cannot prevent crashes of activitylog when the connection to the database is
lost. Then, the duration of the corresponding job will stay empty. I now handle
this situation by checking whether the duration of the last job is set and
suggest possible resolutions to the user. Side-effect: I need to properly
initialise self.lastjob at startup.