bugfixes: regarding duration of last job at startup
To fix this I extended self.lastjob by the duration and now only write the
duration, if lastjob has none.

Then another issue occurred: When you set the start time of a new job and
the last job already has a duration, the duration may make the end time of
the last job be after the start time of the new job. I now check for this
and ask user to resolve.