allow to continue unfinished job at restart of ActivityLog
I added checkLastDuration to handle crashes of ActivityLog when the connection
to the database is lost. Especially, I handled the case that a job needs to be
closed without a directly following job, e.g., the last job of the day. I did
not, however, consider the case that a new job immediately connects to the
last, unfinished job.

I have now changed this. The user can now opt to simply provide a standard job
string to connect to the last job, or the user can continue the last job
(nothing will be done) such that the duration will be written when the next job
is recorded.