bugfix: duration of previous to last job when deleting
When overwriting the last job with a new one and providing a user-defined start
time (e.g.: 'del break @13:45' which overwrites the last job with a break at
13:45) I did not handle that the duration of the previous to last job
may also have to be changed in accordance with the new start time. There are
cases where you want to change the duration of the previous to last job,
especially, when correcting something in the usual command environment, but
there are also cases when you don't want that, e.g., when you correct the first
job of the day. This is why I ask the user to make the decision.