improved error handling and fixing DB integrity errors
The activitylog should now be used with a 'with'-statement. I have added the
corresponding enter and exit functions which, especially, allows me to catch
and print unexpected errors before a shell window closes. To control this
behaviour I have now also added the property 'userclose' to the ActivityLog.

bugfix: DB integrity error occurred when I tried to add a label equal to the
name into the dictionary. I now require the label to be different before it is
added.

bugfix: DB integrity error occurred when I tried to add a label to the
dictionary that was already used by some other entity (may be from other table).
I now check that label does not exist in dictionary before adding into DB.