outsource analysis to new ALogAnalysis module + new weekly projects
ActivityLog now has an instance of ALogAnalysis to run analyses that are
interesting for displaying on the command line while using the activity log.
This is only computing of work hours for now.

Additionally, I started looking into further analyses. Specifically, I created a
function that returns jobs of this week together with the associated projects
and activities answering the question "What have I done this week?"